[SYCL] Re-use OpenCL sampler in SYCL mode
This is an attempt to re-use OpenCL sampler for SYCL.

- `sampler_t` type name is replaced with `__ocl_sampler_t` to avoid
potential collisions with user types
- selectively enabled a couple of OpenCL diagnostics for this type

Notes:
- errors are reported in the SYCL headers rather than in the user's code
- some diagnostics trigger on SYCL use cases. E.g. SYCL kernel
initializes lambda members with the values from kernel arguments. This
initialization code for sampler emits errors because OpenCL disallows
sampler on the right hand side of the binary operators - these are
supposed to be used only by built-in functions. Another potential issue
is the lambda object itself - captured sampler is a member of the lambda
object and OpenCL doesn't allow composite types with samplers. SPIR-V
produced from SYCL should be okay as lambda object can be removed by
standard LLVM transformation passes.

Signed-off-by: Alexey Bader <alexey.bader@intel.com>