[SYCL][AddressSpace] Fix address space of a static variable initializer
When a static variable is lowered from AST to LLVM IR, it can change
an address space depending on its storage class. For example, a static
variable declared in function scope may be assigned to global address
space if language rules allow.

When this happens, original address space of a variable (retuned by
D.getType()) is no longer can be used as target address space. Correct
address space for a global can be obtained from
CodeGenModule::GetGlobalVarAddressSpace function, but in this case we
just copy it from the existing global (OldGV).

Signed-off-by: Andrew Savonichev <andrew.savonichev@intel.com>